Refine and extend data models (#59)
- Remove the schema for reference meteorological dataset metadata
defined within the EYA DEF and instead use the IEA Wind Task 43 WRA Data
Model: #49.
- Move the wind uncertainty assessment from the turbine wind resource
assessment object to the energy assessment object: #35. Also add a field
`wind_speed_to_energy_sensitivity_factor`.
- Clarify in field descriptions that inter-annual variability of plant
performance efficiencies should be included as statistics in the plant
performance assessment datasets, where relevant: #36. It is optional to
include inter-annual variability in the results datasets for individual
categories, sub-categories and sub-category elements. For the overall
plant performance efficiency estimate, the inter-annual variability must
be included.
- Change the term "operational lifetime" to "assessment period" since
the period assessed in the EYA may not be the same as the (expected)
lifetime: #58.
- Do not separate between wind farm ID and reference wind farm ID. The
same ID value (the wind farm ID) can be used for both.
- Add results field to energy assessment schema to cover energy
production results that are not annualised, such as the monthly
production in a specific month.
- Upgrade versions of external packages and tools.
- Upgrade to the new major version of the `erdantic` package and update
the functionality for custom diagram rendering to work with that
version. Also refine and extend functionality for diagram generation.
- Install Graphviz in the GitHub Actions pipeline and include the tests
that were previously omitted since `pygraphviz` could not be installed.