Refine and extend data models for capturing uncertainty statistics (#60)
This PR mainly addresses #53. It introduces the option to define a time
frame and return period for statistics (e.g. the P90 value corresponding
to the first 10 years with a return period of one year). The
descriptions and requirements for gross and net energy assessment
results datasets are refine. It also includes a few minor unrelated
refinements.