Align the turbine model and baseline power mode references with the Power Curve Schema (#66)
This PR addresses issue #56.

We have previously only had a placeholder turbine model identifier in
the turbine configuration data model. This is here updated to reference
the turbine model and baseline power mode with reference to the Power
Curve Schema.

There is an open issue on the Power Curve Schema repo to clarify the
best approach to this referencing, but there has not been a response
yet. I do not see the need to wait for that. We can instead refine this
later if we get suggestions for improvements.

This PR does not include example power curves and associated information
according to the Power Curve Schema. That can be covered in a separate
issue and PR.