Make schema corrections and refinements (#80)
- Fix incorrect use of the `$id` keyword, which closes #78. The `$id`
keyword is now used only in the JSON Schema to define the URI of that
schema, and not in EYA DEF document instances.
- Upgrade external packages and tools to the latest versions.
- Revise the incorrect use of the term return period to instead use the
term evaluation period, which closes #76.
- Align definitions of capacities and capacity factors with updates to
the standard document, which closes #81. Note that we had previously
agreed within the EYA DEF group not to include energy yield assessment
results expressed as capacity factors or full load hours to the data
model, since those values can readily be derived from the energy yield
results and capacity values, without any ambiguity, so would be
duplication of data to include. Given the potential for ambiguity in the
use of reference power for the capacity factor calculation, and since
the capacity factor is included as part of the standard requirements for
the written report, it seems appropriate to include it in the EYA DEF as
optional. When not specified, capacity factor estimates can be derived,
and specified estimates can always be checked against derived estimates.
- Add citation file to the repo, which closes #51.